@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:english_words/english_words.dart';
 
 void main() => runApp(MyApp());
 
@@ -12,7 +13,7 @@ class MyApp extends StatelessWidget {
       // Application theme data, you can set the colors for the application as
       // you want
       theme: ThemeData(
-        primarySwatch: Colors.blue,
+        primarySwatch: Colors.orange,
       ),
       // A widget which will be started on application startup
       home: MyHomePage(title: 'Flutter Demo Home Page'),
@@ -23,7 +24,7 @@ class MyApp extends StatelessWidget {
 class MyHomePage extends StatelessWidget {
   final String title;
 
-  const MyHomePage({@required this.title});
+  const MyHomePage({required this.title});
 
   @override
   Widget build(BuildContext context) {
@@ -32,11 +33,108 @@ class MyHomePage extends StatelessWidget {
         // The title text which will be shown on the action bar
         title: Text(title),
       ),
-      body: Center(
-        child: Text(
-          'Hello, World!',
-        ),
+      body: Column(
+        // crossAxisAlignment: CrossAxisAlignment.center,
+        children: [
+          Image.asset(
+            'images/lake.jpg',
+            width: 640,
+            height: 240,
+            fit: BoxFit.cover,
+          ),
+          Header(),
+          ButtonRow(),
+          DescriptionText(),
+        ],
       ),
     );
+  }
+}
+
+class DescriptionText extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: EdgeInsets.all(32),
+      child: Text(
+        'Even though Helvetica Now has a very healthy range of styles and optical sizes'
+        'to use, pairing it with a serif typeface can extend and transform the tone of'
+        'your design work. The structure, vertical axis, and high stroke contrast of Didot'
+        'LT pairs harmoniously with Helvetica. The formality of Didot LT can lessen the'
+        'blue jeans connotations of Helvetica. Try the eText version of Didot LT which is'
+        'tuned for smaller text and digital work.',
+        softWrap: true,
+      ),
+    );
+  }
+}
+
+class ButtonRow extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    // TODO: implement build
+    return Container(
+      padding: EdgeInsets.all(32),
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+        children: [
+          _buildButtonColumn(Colors.blue, Icons.call, 'CcALL'),
+          _buildButtonColumn(Colors.blue, Icons.near_me, 'ROUTE'),
+          _buildButtonColumn(Colors.blue, Icons.share, 'SHARES'),
+        ],
+      ),
+    );
+  }
+
+  Column _buildButtonColumn(Color color, IconData icon, String label) {
+    return Column(
+      children: [
+        Icon(icon, color: color),
+        Container(
+          margin: const EdgeInsets.only(top: 8),
+          child: Text(
+            label,
+            style: TextStyle(
+              fontSize: 12,
+              fontWeight: FontWeight.w400,
+              color: color,
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+}
+
+class Header extends StatelessWidget {
+  const Header({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+        padding: EdgeInsets.all(32),
+        child: Row(
+          children: [
+            Expanded(
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Text(
+                    "Interesting mountain range",
+                    style: TextStyle(fontWeight: FontWeight.bold),
+                  ),
+                  Text("Please join me in the awesome adventure"),
+                ],
+              ),
+            ),
+            Icon(
+              Icons.star,
+              color: Colors.red[500],
+            ),
+            Text('44'),
+          ],
+        ));
   }
 }
